pyproject.toml
[tool.poetry]
name = "gyst"
version = "0.1.0"
description = "Habit Tracker Api"
authors = ["Mariusz Dąbal <mariusz9dabal@gmail.com>", "Dobromir Matuszak <matuszak.d92@gmail.com>", "Maciej Łatosz <Amacedmace@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
Django = "^4.1.5"
django-cors-headers = "^3.13.0"
mysqlclient = "^2.1.1"
django-colorfield = "^0.8.0"
django-dotenv = "^1.4.2"
black = "^22.12.0"
djangorestframework-simplejwt = "^5.2.2"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
django = "^4.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
